Improve header menu
